+++ Changelog.md
@@ -5,8 +5,12 @@
 1. Fix restoring submesh material transparency
     * [Pull request #2536](https://bitbucket.org/osrf/gazebo/pull-request/2536)
 
+1. Updated `gz_log` tool to use `ignition::math`.
+    * [Pull request #2532](https://bitbucket.org/osrf/gazebo/pull-request/2532)
+
 1. Update plugins to use ign-math.
     * [Pull request #2531](https://bitbucket.org/osrf/gazebo/pull-request/2531)
+    * [Pull request #2534](https://bitbucket.org/osrf/gazebo/pull-request/2534)
 
 1. Use ignition math with `rendering/Distortion` and update function names.
     * [Pull request #2529](https://bitbucket.org/osrf/gazebo/pull-request/2529)
@@ -119,6 +123,7 @@
     * [Pull request #2514](https://bitbucket.org/osrf/gazebo/pull-request/2514)
     * [Pull request #2522](https://bitbucket.org/osrf/gazebo/pull-request/2522)
     * [Pull request #2525](https://bitbucket.org/osrf/gazebo/pull-request/2525)
+    * [Pull request #2533](https://bitbucket.org/osrf/gazebo/pull-request/2533)
 
 1. Add Wind support
     * [Pull request #1985](https://bitbucket.org/osrf/gazebo/pull-request/1985)